requirements.txt

djangorestframework
djangorestframework-simplejwt
django
django-cors-headers

gunicorn


